lab 2.1

+++ LAB1/cpp_files/list_of_orders_class.cpp
@@ -1,7 +1,7 @@
 #include "../header_files/list_of_orders_class.h"
 #include "../header_files/visitor_class.h"
 
-void List_of_orders:: add_data(const string& v_name,const string& ph_number,const vector<string>& items,float f_cost)
+void List_of_orders:: add_data(const string& v_name,const string& ph_number,const vector<string>& items,double f_cost)
 {
     orders.emplace_back(v_name,ph_number,items,f_cost);
 }
